+++ frontend/src/views/components/write/Write.css
@@ -68,7 +68,7 @@
   right: 60px;
   width: 15px;
   height: 15px;
-  background: url(../../../imgs/icons/plus.svg);
+  background: url(../../../imgs/icons/plus.svg) no-repeat;
   background-size: 15px 15px;
 }
 
@@ -142,4 +142,18 @@
 .clrInput {
   width: 80px;
   height: 40px;
+}
+
+.postTextWrite {
+  box-sizing: border-box;
+  padding: 15px 0 15px 0;
+  border: 0;
+  outline: 0;
+  background: #f0f3f6;
+  border-radius: 12.5px;
+  text-align: center;
+  font-size: 13px;
+  margin-top: 10px;
+  width: 250px;
+  height: 50px;
 }
